Implements mock MinioHandler
Implements mock MinioHandler, which allows for testing without
polluting the Minio store.
